Link Zoonotic TB Relevant Data

Link data relevant to zoonotic TB data. Data is linking process maximises preserved data so any analysis with this data may need additional exclusions and/or munging.

link_data(z_tb_humans = NULL, tb_humans = NULL, z_tb_animals = NULL,
  demo = NULL, animal_demo = NULL, verbose = TRUE)

Arguments

z_tb_humans
Data on zoonotic TB in humans - formatted as EstZoonoticTB::zoonotic_tb_humans. Defaults to EstZoonoticTB::zoonotic_tb_humans.

tb_humans
Data on TB in humans - formatted as EstZoonoticTB::tb_data(). Defaults to EstZoonoticTB::tb_data().

z_tb_animals
Data on zoonotic TB in animals - formatted as EstZoonoticTB::zoonotic_tb_animals. Defaults to EstZoonoticTB::zoonotic_tb_humans.

demo
Data on demographics - formatted as EstZoonoticTB::demographics. Defaults to EstZoonoticTB::demographics.

animal_demo
Data on demographics in animals - formatted as EstZoonoticTB::animal_demographics. Defaults to EstZoonoticTB::animal_demographics.

verbose
Logical, defaults to TRUE. Should verbose progress messages be shown.

Value

A dataframe of linked data containing all available global data relevant to zTB in humans.

Examples

## Linked data relying on package defaults
df <- link_data()

#> Joining human and animal demographic data.

#> Countries with data present for demographics and not animal demographics:

#>  [1] "Andorra"                  "Anguilla"                
#>  [3] "Aruba"                    "Channel Islands"         
#>  [5] "China, Macao SAR"         "Gibraltar"               
#>  [7] "Holy See"                 "Isle of Man"             
#>  [9] "Kiribati"                 "Maldives"                
#> [11] "Marshall Islands"         "Mayotte"                 
#> [13] "Monaco"                   "Nauru"                   
#> [15] "Northern Mariana Islands" "Palau"                   
#> [17] "San Marino"               "Tokelau"                 
#> [19] "Turks and Caicos Islands" "Tuvalu"                  

#> Countries with data present for animal demographics and not demographics:

#> [1] "Norfolk Island"

#> Joining TB incidence in humans data and zTB presence in animals data using country codes.

#> Countries with data present for TB incidence and not zTB in animals:

#>  [1] "American Samoa"                       
#>  [2] "Anguilla"                             
#>  [3] "Antigua and Barbuda"                  
#>  [4] "Aruba"                                
#>  [5] "Bahamas"                              
#>  [6] "Benin"                                
#>  [7] "Bermuda"                              
#>  [8] "Bonaire, Saint Eustatius and Saba"    
#>  [9] "British Virgin Islands"               
#> [10] "Cameroon"                             
#> [11] "Cayman Islands"                       
#> [12] "China, Macao SAR"                     
#> [13] "Curaçao"                              
#> [14] "Democratic People's Republic of Korea"
#> [15] "Dominica"                             
#> [16] "Equatorial Guinea"                    
#> [17] "Gabon"                                
#> [18] "Grenada"                              
#> [19] "Guam"                                 
#> [20] "Guatemala"                            
#> [21] "Lebanon"                              
#> [22] "Luxembourg"                           
#> [23] "Maldives"                             
#> [24] "Monaco"                               
#> [25] "Montserrat"                           
#> [26] "Nauru"                                
#> [27] "Niue"                                 
#> [28] "Northern Mariana Islands"             
#> [29] "Puerto Rico"                          
#> [30] "Rwanda"                               
#> [31] "Saint Kitts and Nevis"                
#> [32] "Seychelles"                           
#> [33] "Sint Maarten (Dutch part)"            
#> [34] "Solomon Islands"                      
#> [35] "Tajikistan"                           
#> [36] "Togo"                                 
#> [37] "Tokelau"                              
#> [38] "Turks and Caicos Islands"             
#> [39] "Tuvalu"                               
#> [40] "Wallis and Futuna Islands"            
#> [41] "Yemen"                                
#> [42] "Netherlands Antilles"                 
#> [43] "Serbia & Montenegro"                  

#> Countries with data present for zTB in animals and not TB incidence:

#>  [1] "Ceuta"                       "Chinese Taipei"             
#>  [3] "Falkland Islands (Malvinas)" "Faroe Islands"              
#>  [5] "French Guiana"               "Guadeloupe (France)"        
#>  [7] "Liechtenstein"               "Martinique"                 
#>  [9] "Mayotte (France)"            "Melilla"                    
#> [11] "Reunion"                     "St. Helena"                 

#> Joining zTB incidence in humans data and all other TB data using country names

#> Joining TB data and demographic data using country names.

#> Countries with data present for TB not for demographics:

#>  [1] "Bonaire, Saint Eustatius and Saba" "Curaçao"                          
#>  [3] "Sint Maarten (Dutch part)"         "West Bank and Gaza Strip"         
#>  [5] "Netherlands Antilles"              "Serbia & Montenegro"              
#>  [7] "Ceuta"                             "Chinese Taipei"                   
#>  [9] "Guadeloupe (France)"               "Mayotte (France)"                 
#> [11] "Melilla"                           "St. Helena"                       

#> Countries with data present demographics and not TB:

#> (Some mismatches are to be expected here due to historic country names)

#>  [1] "Belgium-Luxembourg"              "Channel Islands"                
#>  [3] "Czechoslovakia"                  "Ethiopia PDR"                   
#>  [5] "Gibraltar"                       "Guadeloupe"                     
#>  [7] "Holy See"                        "Isle of Man"                    
#>  [9] "Mayotte"                         NA                               
#> [11] "Pacific Islands Trust Territory" "Palestine"                      
#> [13] "Saint Pierre and Miquelon"       "Sudan (former)"                 
#> [15] "United States Virgin Islands"    "USSR"                           
#> [17] "Western Sahara"                  "Yugoslav SFR"                   
#> [19] "Norfolk Island"                 

df

#> # A tibble: 15,465 x 26
#>    country country_code g_whoregion  year tb_cases tb_inc tb_inc_lo tb_inc_hi
#>    <fct>   <fct>        <fct>       <dbl>    <int>  <dbl>     <dbl>     <dbl>
#>  1 Afghan… AFG          Eastern Me…  2000    39000    190       123       271
#>  2 Afghan… AFG          Eastern Me…  2001    41000    189       123       271
#>  3 Afghan… AFG          Eastern Me…  2002    43000    189       122       270
#>  4 Afghan… AFG          Eastern Me…  2003    45000    189       122       270
#>  5 Afghan… AFG          Eastern Me…  2004    47000    189       122       270
#>  6 Afghan… AFG          Eastern Me…  2005    48000    189       122       270
#>  7 Afghan… AFG          Eastern Me…  2006    50000    189       122       270
#>  8 Afghan… AFG          Eastern Me…  2007    51000    189       122       270
#>  9 Afghan… AFG          Eastern Me…  2008    52000    189       122       270
#> 10 Afghan… AFG          Eastern Me…  2009    54000    189       123       270
#> # … with 15,455 more rows, and 18 more variables: prop_tb_ep <dbl>,
#> #   prop_hiv <dbl>, prop_hiv_lo <dbl>, prop_hiv_hi <dbl>,
#> #   z_tb_dom_animal <fct>, z_tb_wild_animal <fct>, z_tb_id <int>,
#> #   z_tb_geo_coverage <fct>, z_tb_study_pop <fct>, z_tb_multi_year_study <fct>,
#> #   tb_z_prop <dbl>, tb_z_prop_lo <dbl>, tb_z_prop_hi <dbl>,
#> #   tb_z_prop_se <dbl>, population <dbl>, prop_rural <dbl>, cattle <int>,
#> #   cattle_per_head <dbl>